+++ kcommons/src/main/java/com/kairos/dto/user/country/agreement/cta/cta_response/DayTypeDTO.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.validation.constraints.NotBlank;
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -18,7 +19,7 @@ import java.util.Map;
 @NoArgsConstructor
 public class DayTypeDTO {
 
-    private Long id;
+    private BigInteger id;
     @NotBlank(message = "error.DayType.name.notEmpty")
     private String name;
     private List<Day> validDays = new ArrayList<>();
@@ -34,7 +35,7 @@ public class DayTypeDTO {
     private Map<String,String> translatedDescriptions;
     private Map<String, TranslationInfo> translations;
 
-    public DayTypeDTO(Long id, String name, List<Day> validDays, List<CountryHolidayCalenderDTO> countryHolidayCalenderData, boolean holidayType, boolean allowTimeSettings) {
+    public DayTypeDTO(BigInteger id, String name, List<Day> validDays, List<CountryHolidayCalenderDTO> countryHolidayCalenderData, boolean holidayType, boolean allowTimeSettings) {
         this.id = id;
         this.name = name;
         this.validDays = validDays;
@@ -43,7 +44,7 @@ public class DayTypeDTO {
         this.allowTimeSettings = allowTimeSettings;
     }
 
-    public DayTypeDTO(Long id, String name, List<Day> validDays, List<CountryHolidayCalenderDTO> countryHolidayCalenderData, boolean holidayType, boolean allowTimeSettings,String colorCode) {
+    public DayTypeDTO(BigInteger id, String name, List<Day> validDays, List<CountryHolidayCalenderDTO> countryHolidayCalenderData, boolean holidayType, boolean allowTimeSettings,String colorCode) {
         this.id = id;
         this.name = name;
         this.validDays = validDays;
